feat: post-page 생성(예시 파일은 추후에 삭제)
